+++ internal/integration_test/fuzz/wazerolib/nodiff_test.go
@@ -7,8 +7,8 @@ import (
 	"github.com/tetratelabs/wazero/internal/testing/require"
 )
 
-// TestReRunFailedCase re-runs the failed case specified by WASM_BINARY_NAME in testdata directory.
-func TestReRunFailedCase(t *testing.T) {
+// TestReRunFailedRequireNoDiffCase re-runs the failed case specified by WASM_BINARY_NAME in testdata directory.
+func TestReRunFailedRequireNoDiffCase(t *testing.T) {
 	binaryPath := os.Getenv("WASM_BINARY_PATH")
 
 	wasmBin, err := os.ReadFile(binaryPath)
