+++ internal/integration_test/fuzz/wazerolib/validate_test.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"os"
+	"testing"
+)
+
+// TestReRunFailedValidateCase re-runs the failed case specified by WASM_BINARY_NAME in testdata directory.
+func TestReRunFailedValidateCase(t *testing.T) {
+	binaryPath := os.Getenv("WASM_BINARY_PATH")
+
+	wasmBin, err := os.ReadFile(binaryPath)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	tryCompile(wasmBin)
+}
